+++ src/components/TextList/index.ts
@@ -1,0 +1,1 @@
+export { default as TextList } from './TextList';
